requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
datasets==2.20.0
dill==0.3.8
diskcache==5.6.3
evaluate==0.4.2
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.5.0
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
huggingface-hub==0.24.5
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
llama_cpp_python==0.2.85
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
multiprocess==0.70.16
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
packaging==24.1
pandas==2.2.2
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
pyarrow==17.0.0
pyarrow-hotfix==0.6
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
regex==2024.7.24
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
safetensors==0.4.3
six==1.16.0
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
-e git+https://github.com/huggingface/transformers.git@c1aa0edb48217f416f4bbe6e3a9db1500284513b#egg=transformers
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
Werkzeug==3.0.3
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
